SystemEditor: add/remove bodies, cleanup
- Split *EditorAPI classes into separate file
- Add new file for Undo helper classes
- Implement creation and deletion of non-root SystemBodies.
- Further work required to set sane/sensible defaults for new body and handle making a body the new root.
